move plugins to `plugins.py` #419
